feat(core-auth): move to core root
