+++ packages/core/auth/index.ts
@@ -4,5 +4,5 @@ export { XmAuthenticationService} from './src/xm-authentication.service';
 export { XmAuthenticationStoreService } from './src/xm-authentication-store.service';
 export { XmCoreAuthModule } from './src/xm-core-auth.module';
 export { XmUserService } from './src/xm-user.service';
-export { XmSessionService } from '../xm-session.service';
+export { XmSessionService } from '../src/xm-session.service';
 export { XmUser, XmUserLogin, XmUserPermission } from './src/xm-user-model';
