+++ packages/core/auth/package.json
