+++ packages/core/auth/src/xm-core-auth.module.ts
