+++ src/app/shared/auth/auth-jwt.service.ts
@@ -5,8 +5,8 @@ import { XmSessionService } from '@xm-ngx/core';
 import { SessionStorageService } from 'ngx-webstorage';
 import { Observable } from 'rxjs';
 import { map, switchMap, tap } from 'rxjs/operators';
-import { AuthRefreshTokenService } from '../../../../packages/core/src/auth/src/auth-refresh-token.service';
-import { XmAuthenticationStoreService } from '../../../../packages/core/src/auth/src/xm-authentication-store.service';
+import { AuthRefreshTokenService } from './auth-refresh-token.service';
+import { XmAuthenticationStoreService } from './xm-authentication-store.service';
 
 import { DEFAULT_AUTH_TOKEN, DEFAULT_CONTENT_TYPE } from '../../xm.constants';
 import { CustomUriEncoder } from '../helpers/custom-uri-encoder';
