+++ tsconfig.json
@@ -24,7 +24,7 @@
       "@xm-ngx/components/*": ["packages/components/src/*"],
       "@xm-ngx/core": ["packages/core"],
       "@xm-ngx/core/theme": ["packages/xm-theme/public_api"],
-      "@xm-ngx/core/auth": ["packages/core/src/auth"],
+      "@xm-ngx/core/auth": ["packages/core/auth"],
       "@xm-ngx/ext/*": ["src/app/ext/*"],
       "@xm-ngx/core/environment": ["src/environments"],
       "@xm-ngx/core/permission": ["src/app/shared/privilege"],
